Move EndpointAnalytics out of EndpointMuxer into its own EndpointMuxerAnalytics file
